Refactor OpalDefendantAccountHistoryService to integrate PaymentTermsHistorySourceService for fetching payment terms
